The PPU should handle displaying the background, but I have no way to test that
I'm thinking the best way to deliver the frame to display would be every cycle the emulator polls the PPU to see if it is ready. More of a pull observer model. We could also declare an observer interface and push it down.

I'm handling the PPU states internally, it seems more straightforward to me that way.