PPU is done but untested. FIFO is done, sprites are done, window done, the works.
I created an IEmulator (feel free to change) with my needed interactions. Mainly trigger interrupts and pass the screen data.

All the Emulator needs to do with the PPU is tick it. The DMA also needs to be ticked. State information is handled internally.

With this I am signing off from coding until we have a limping CPU that can populate VRAM and OAMRAM.